new fullscreen navbar hiding script
we already had some CSS to hide the navbar in fullscreen and reveal it when hovering. but it didn't interact well with the various popups in the UI. i wrote a new script that reveals the navbar when navbar-related popups are open, regardless of hovering/focusing, and also reveals the navbar when it's being hovered or the urlbar is focused. it wouldn't be possible without CSS since you're limited to using attributes and pseudo-selectors like :hover or [open="true"] on high-level elements with the navbar box as their subsequent sibling (using ~ combinator). and most of the popups are children of mainpopupset and lack any DOM attributes to tell you they're open. so we listen for popups showing and hiding and use that to set our own attributes on navigator-toolbox. and separately from that, use a mutation observer to listen for the urlbar being focused, and use that to set another attribute on navigator-toolbox (which is far from a subsequent sibling of the urlbar). we can't use an ordinary onfocus event listener for the urlbar because there are various features in firefox that open the urlbar without technically focusing it. like there's the universal focus concept and then there's also an original attribute that firefox uses for the urlbar, "focused." since that attribute is separate from the universal "focus" and "blur" values, this attribute can be set to true without actually focusing the urlbar. so we have to listen to the attribute itself and the only way i know of to do that is a mutation observer. i don't think it is much more costly in terms of computation, because we can filter out any other types of mutations and any other attributes besides focused. this script requires some of the CSS rules in uc8 to work. search uc8 for "fullscreenNavBar" to find the section. the exact formatting of the elements that get adjusted by this sizemode modulation is specified in other files so if you're cherry picking stuff from this repo you'll probably have to do some tweaking to the navbar, findbar, and sidebar box.