new toolbar overflow slider
overhauled the toolbar overflow slider so it now uses the mousewheel scrolling behavior of firefox's tab bar. favorite thing i've made yet, so simple yet so nice! it even blocks scrolling input when a toolbar popup is open. vertical mousewheel scrolls the toolbar horizontally. put all the toolbar items you wanna wrap to the right of the urlbar.  so to the left should be back button, fwd button, reload, and urlbar/searchbar. everything else to the right. remove everything from the overflow menu and hide the overflow widget with css if you want. theres also a script that does the same sort of thing for the search one-offs, but it requires a lot of supporting CSS to reposition the one-offs so they're "inside" the urlbar. unfortunately we can't use javascript to actually put the one-offs into the urlbar input area without royally fucking things up. so to use the scrollingoneoffs.uc.js you'll have to rifle through my css files to find the stuff you need. just ctrl+f for "one-off"