about:userchrome 1.1.4
Account for possibility of restarting Firefox on about:userchrome with a
script expanded that was deleted between restarts. Previously if this
happened, the page would be blank and you'd have to click the back
button to get back to the scripts page, since it was open to a card for
a nonexistent script. Now if the script doesn't exist, it will
automatically navigate back to the scripts main page.