Fix PDF highlighting and update context menu icons
Add context menu icon for Format with Prettier extension.

Adjust context menu icons for search engines: contextMenuMods.uc.js adds
a context menu for search engines with icons. I use all my own icons for
search engines elsewhere in the browser. I replace entire search engines
(my versions are in the readme) where possible, but the Google and
DuckDuckGo engines can't be replaced cleanly, so we have to use the
versions bundled with Firefox. But they have color icons that don't
really match the theme, which we can't directly change, so we replace
every instance of them in CSS. I had neglected to replace them in this
particular context menu, however. This update fixes that.