style: adopt more typical rounded tabs
I have put off doing this for a long time since I was very attached to
the folder-style tabs, but it has been causing problems for a long time.
It forced me to let tabs flex to the full width of the screen if they
were the only tab, since showing and hiding the rounded outer corners
when tabs overflow would cause an infinite loop of overflowing and
underflowing. It also caused somewhat wonky animations. The animation
problems got really bad with recent Firefox changes, so I finally bit
the bullet and brought the theme back in line with vanilla Firefox.