feat: support tab hover previews!
tabMods.uc.js now supports tab hover previews. the delay was because the
code was constantly changing, making it hard to maintain a single
tooltip for the tab, the close button, and the sound button. until now,
managing all three tooltips was handled by restoreTabSoundButton. but
now, they are mainly handled in tabMods. you still need
restoreTabSoundButton, since it adds styles for the sound button. but
now you can enable previews with `browser.tabs.hoverPreview.enabled` if
you want. if you disable that pref, it will continue using the custom
tooltip from restoreTabSoundButton. if you enable it, it will use the
tab preview panel from tabMods. so you need both files, and it is the
pref that decides which type of tooltip you get. various styles have
also been added to refine the tab preview panel.